rename lib to js
