+++ src/js/editorCopy.js
