+++ src/js/markerSettingsGet.js
