+++ src/js/permissions.js
