+++ src/js/permissionsCheck.js
