+++ src/js/permissionsToggle.js
