+++ src/js/permissionsWatch.js
